Conform to MVC.
Users will no longer access the JSP files directly. Instead they will access the servlets directly via the url mapping defined in web.xml. The doGet method was modified to internally (server-side) redirect the user to the appropriate JSP page. 

New URLS:
/login.jsp -> /login
/register.jsp -> /register
/products.jsp -> /shop
/user.jsp -> /user
/index.jsp -> /

By doing it this way, we remove the need of scriptlets in JSP pages (which is highly discouraged in the JSP world) and instead set attributes within the servlet (controller) and access the attributes from the JSP page (view).