Implement update cart and add to cart quantity.
Adds the ability to add to cart with a quantity.
Adds the ability to update the cart on the cart page. If the quantity of an item is set to 0 in the cart, it will remove the item from the cart.
The cart subtotals are now updated.
Cart icon on nav bar now correctly shows the number of items in the cart.
Duplicate items are now shown as one item with the amount of that item in that cart.